openeo_argoworkflows/executor/pyproject.toml
[tool.poetry]
name = "openeo-argoworkflows-executor"
version = "2024.6.1"
description = ""
authors = ["sean <sean.hoyal@external.eodc.eu>"]
readme = "README.md"
packages = [{include = "openeo_argoworkflows_executor"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
dask-gateway = "^2024.1.0"
openeo-processes-dask = {version = "2024.5.2", extras = ["implementations", "ml"]}
openeo-pg-parser-networkx = ">=2024.4.0"
pystac-client = ">=0.8.2,<1"
rioxarray = ">=0.15.1,<1"
xarray = ">=2024.6.0"
click = ">=8.1.7"
pydantic = "^2.7.0"
stactools = "^0.4.8"
odc-stac = ">=0.3.3,<1"
haversine = "^2.8.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
ipykernel = "^6.15.1"
pre-commit = "^2.20.0"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.23.0"
pydoc-markdown = "4.8.2"

[tool.poetry.scripts]
openeo_executor = "openeo_argoworkflows_executor.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
